refactor(119015): :hammer: cria processador de planilha generico
